Serialization - merging vectors when loading in APIMapping, working NEED refactoring.
When loading json array into vector, if the vector contain and element
with the same id as one of the json array objects we merge the object
and the array element otherwise we delete the array element.

if the json array have value that isn't in the vector (not in the vector
= dont have the same id) we add a new element to the vector.

- The code is written REALLY poorly, need to refactor it (simple in most
parts) i just wanted to finish the use case of loading from/into the GUI
so I quick and dirty it.