Solved the space problem (for now) + Rearranged the files
Solved the space problem:
- constant string are in the heap, that plus temlates made that they got
multiplied. i used the macro to put them in the flash.
- __PRETTY_FUNCTION__ - the macro to put it in the flash didnt work so
got read of it and the id now is the pointer to the (specific to that
template) function

Rearranged the files:
makes a little more sense, libs and frameworks in one place while the
code which uses it in another